requirements.txt
faster-whisper
google-generativeai
python-dotenv
yt-dlp>=2024.10.0
torch>=2.0.0
torchaudio>=2.0.0
requests>=2.31.0
python-dateutil>=2.8.2
openpyxl
pandas
sqlalchemy
pymysql
cryptography
